get lambda function code ready to talk to DynamoDB table
