+++ aws-backend/nft-options/create.js
@@ -4,23 +4,40 @@ const AWS = require('aws-sdk'); // eslint-disable-line import/no-extraneous-depe
 
 const dynamoDb = new AWS.DynamoDB.DocumentClient();
 
-module.exports.get = (event, context, callback) => {
+module.exports.create = (event, context, callback) => {
+  const timestamp = new Date().getTime();
+  const data = JSON.parse(event.body);
+  
+  
+
   const params = {
     TableName: process.env.DYNAMODB_TABLE,
-    Key: {
-      id: event.pathParameters.id,
+    Item: {
+      StrikePrice: 111,
+      Premium: 111,
+      Signature: 111,
+      Duration: 222,
+      OwnerAccountAddress: data.Owner,
+      Hash: "test123",
+      NftContractAddress: data.NftContractAddress
+      
     },
   };
 
-  // fetch todo from the database
-  dynamoDb.get(params, (error, result) => {
+
+
+
+
+
+  // write the todo to the database
+  dynamoDb.put(params, (error) => {
     // handle potential errors
     if (error) {
       console.error(error);
       callback(null, {
         statusCode: error.statusCode || 501,
         headers: { 'Content-Type': 'text/plain' },
-        body: 'Couldn\'t fetch the todo item.',
+        body: 'Couldn\'t create the item.',
       });
       return;
     }
@@ -28,7 +45,7 @@ module.exports.get = (event, context, callback) => {
     // create a response
     const response = {
       statusCode: 200,
-      body: JSON.stringify(result.Item),
+      body: JSON.stringify(params.Item),
     };
     callback(null, response);
   });
