+++ aws-backend/nft-options/delete.js
