+++ aws-backend/notes.md
@@ -1,0 +1,23 @@
+
+
+
+## Curl query for post
+
+curl -X POST -H "Content-Type: application/json" \
+    -d '{"Hash": "linuxize", "NftContractAddress": "123ab", "Owner": "0xcde123", "Duration": 1 , "Signature": "vvvv", "Premium":1111, "StrikePrice": 1111}' \
+    https://6bt88e77ie.execute-api.us-east-1.amazonaws.com/dev/nft-options/
+
+
+
+## Curl query for post
+
+curl -X GET  \                                    
+    https://6bt88e77ie.execute-api.us-east-1.amazonaws.com/dev/nft-options/0x222221111cde1sdssdsdsdsdsdsdsdsdsd2sss3
+
+
+
+
+## Query Builder
+
+https://dynobase.dev/dynamodb-query-builder/
+
