+++ aws-backend/serverless.yml
@@ -1,10 +1,9 @@
 service: serverless-rest-api-with-dynamodb
 
-frameworkVersion: ">=1.1.0 <2.0.0"
-
+frameworkVersion: '2'
 provider:
   name: aws
-  runtime: nodejs10.x
+  runtime: nodejs14.x
   environment:
     DYNAMODB_TABLE: ${self:service}-${opt:stage, self:provider.stage}
   iamRoleStatements:
@@ -16,64 +15,48 @@ provider:
         - dynamodb:PutItem
         - dynamodb:UpdateItem
         - dynamodb:DeleteItem
-      Resource: "arn:aws:dynamodb:${opt:region, self:provider.region}:*:table/${self:provider.environment.DYNAMODB_TABLE}"
+      #todo: tighten later
+      Resource: "arn:aws:dynamodb:${opt:region, self:provider.region}:*:table/*"
 
 functions:
   create:
-    handler: todos/create.create
+    handler: nft-options/create.create
     events:
       - http:
-          path: todos
+          path: nft-options
           method: post
           cors: true
 
   list:
-    handler: todos/list.list
+    handler: nft-options/list.list
     events:
       - http:
-          path: todos
+          path: nft-options
           method: get
           cors: true
 
   get:
-    handler: todos/get.get
+    handler: nft-options/get.get
     events:
       - http:
-          path: todos/{id}
+          path: nft-options/{id}
           method: get
           cors: true
 
   update:
-    handler: todos/update.update
+    handler: nft-options/update.update
     events:
       - http:
-          path: todos/{id}
+          path: nft-options/{id}
           method: put
           cors: true
 
   delete:
-    handler: todos/delete.delete
+    handler: nft-options/delete.delete
     events:
       - http:
-          path: todos/{id}
+          path: nft-options/{id}
           method: delete
           cors: true
 
-resources:
-  Resources:
-    TodosDynamoDbTable:
-      Type: 'AWS::DynamoDB::Table'
-      DeletionPolicy: Retain
-      Properties:
-        AttributeDefinitions:
-          -
-            AttributeName: id
-            AttributeType: S
-        KeySchema:
-          -
-            AttributeName: id
-            KeyType: HASH
-        ProvisionedThroughput:
-          ReadCapacityUnits: 1
-          WriteCapacityUnits: 1
-        TableName: ${self:provider.environment.DYNAMODB_TABLE}
+
