+++ aws-backend/todo.md
@@ -1,0 +1,4 @@
+# TODO on AWS 
+
+- parameterize DynamoDB table as Lambda function environment variable. 
+- find a way to work with multiple serverless.yml file without destroying other stack
